analizer/pyproject.toml
[tool.poetry]
name = "almaz-test"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
httpx = "^0.27.2"
flower = "^2.0.1"
loguru = "^0.7.2"
alembic = "^1.14.0"
pytest = "^8.3.3"
prometheus-fastapi-instrumentator = "^7.0.0"
lxml = "^5.3.0"
celery = {extras = ["librabbitmq", "sqlalchemy"], version = "^5.4.0"}
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"
flake8 = "^7.1.1"


[tool.pytest.ini_options]
markers = [
    "parser",
    "prompt",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
